factor out runner from juju-core.
We make a few changes when moving from juju-core.
In particular, there's no particular use for NewRunner to
return an interface rather than a concrete type, so we
return a concrete type like a conventional Go package.

We also make the parameters to NewRunner a struct so that we won't need
to break compatibility if we want to change them.

We change the tests so we no longer have a dependency on
github.com/juju/juju and simplify them a little.