Treat any error as potential for exponential backoff.
Rather than only backing off when Start fails, use a timeout for any
error within a window of starting ends up being treated as a potential
for exponential backoff.
We need to decide what a reasonable length of time is before we decide
that the worker was effectively working-as-normal and should be
restarted immediately.