Synchronize the test with the processing loop.
With the previous code, there was a race between when the Manifold saw
the start event, and when it would emit "started" back to the core loop,
and the core loop would track the start time. This meant that the test
suite could see the Manifold observe the Started event, and then we
would move the clock forward in time *before* the engine would finish
observing the started event and record the start time.
By using an Install, we know the worker has finished processing the
Start and it is safe to move time forward.

This is done by running:
  go test -race -failfast -count=200 -check.v -check.f Backoff

Which would fail about 1-in-20 times, but doesn't fail after 200
attempts.