Metrics: Allow the collection of worker lifecycle events
The following starts to expose the events of a worker, so that we can
better observe the changes of a juju instance. The only one that seems
relevant at the moment, is how many times a start is done, but this can
easily be expanded at a later date.

I chose RecordStart to be explicit, rather than picking Record with a
state. Although changing that in the future should be really easy.