Updates worker package to latest juju errors.
We are updating the worker package to take advantage of new changes
available in the Juju errors package. Specifically this affords
ConstError, Is and As support.

We want this so packages that rely on this one such as Juju can pass
heavily wrapped errors and still have the same side effect in this
package.

The example here is that at the moment Juju can't wrap error's from this
package such as uninstall as this library won't handle the case
correctly as the error gets lost in Cause.