Catacomb context support
Adds context support to the catacomb type. In reality this just calls
the tomb context method. By exposing the context method we can then tie
a context to the lifecycle of a catacomb.

The only difference between a tomb and a catacomb in relation to a
context is that a catacomb doesn't expose the concept of a parent
context. The tomb type allows you to create a tomb from a context
(WithContext), which will be the root context for the tomb. Instead a
catacomb will currently fallback to a context.Background. We could
provide a WithContext for a catacomb, I'm unsure if that functionality
is required right now?

---

The tests were lifted and modified from the tomb test package, which
ensures that we don't deviate from the tomb package in the future.