Prevent start worker whilst dying
The start worker method only checks to see if the tomb is dead in
the select statement. This has the potential of doing additional
work of attempting to create a worker whilst it is dying.

By changing this to tomb dying state, this prevents sending the
start request to the start channel. This should reduce the busy
work that a runner will do once it starts to shutdown. With the
benifit of stopping the runner in a quickly, as the runner no
longer need to process each request whilst dying.