Expose ErrDying when dying
The following ensurse that we return ErrDying if we fail to start
the worker when we're dying. In addition, because tomb errors are
sequenced from ErrDying to ErrDead. All ErrDeads are also ErrDying
states. To ensure that we send ErrDead in a reliable way, we recheck
the Dead state of the tomb when dying. This ensures that we get the
right error message reliably.