Scoped context
A scoped context is a context.Context that is bound to a dependency
context. If the dependency context is aborted, then the context.Context
is also done.

This will provide a nice way to pass a context into workers and watchers
from the engine. In an ideal world each dependency.Context is also a
context.Context, but that will require way too many changes and potential
breakages.