Dependency context.Context introduction
The following reworks the dependency engine to bring in a real
context.Context. The fallout of this means that we decouple the
concept of cancelation and where you get the dependency from. In
doing so, context has become snapshot. The original comments
around this type just state that it is a snapshot, so we might as
well just call it that.
The dependency Context becomes split. Removing the cross cutting
concerns. The aborting of a worker in the start now becomes
context.Context and the retrival of a dependency is now a Container.
Currently we're just creating a adhoc context.Context, but in the
next commit, we'll tie this to the tomb directly, ensuring that upon
destruction of the engine, all manifolds will be killed.