Scoped context
The context creation is now tied to the lifecycle of the tomb. Doing
this ensures that when the engine is killed, then the context is also
killed. It worth noting that cancel is idempotent so, calling it multiple
times does nothing. As the main loop on tomb death is also attempting
to cancel the context of the snapshot worker, it essentially becomes a
no-op.