Expose the worker names
The following exposes the worker names for a given runner. This is
useful to iterate over the workers. This doesn't check to see if
the workers are not dead/found, it just reports the actual workers
available to it.

This rationale for this, is to prevent the divergent sources of truth.
The runner knows exactly which runners it is tracking and removes them
after death. Attempting to track when each runner is dead from the
outside isn't that easy, if some can crash internally or be removed
from the runner directly. Instead of playing a guessing game, just
look it up.