Goroutine profile labels
Adds the runtime.Do function around certain goroutines to help with
debugging which goroutines are the potential bottlenecks. Juju uses
the runner and catacomb types as basic building blocks, to then
build larger types with, it becomes hard to diagnose which specific
worker or goroutine is becoming the problem.

Using the an old post from PolarSignals[1] it's possible to assoicate
a label to the lifetime of a goroutine. This has been available since
2016, so is safe to assume it's stable and useful.

The new function call is added to two areas:

 - Runner. Adds pprof.Do to the start and restart of the startFunc. The
   labels for the call is then tracked in the worker info. This ensures
   that the labels are consistent between restarts.
 - Catacomb. This adds the pprof.Do to only the work that the catacomb
   supervises directly (Plan.Work). The slice of workers in Plan.Init,
   can be added later on if we find it useful.

In order to add useful names to the labels, we unfortunately have to
use reflect to gain package names. This keeps the worker.Worker interface
the same, preventing a breaking change, for debugging purposes only.

 1. https://www.polarsignals.com/blog/posts/2021/04/13/demystifying-pprof-labels-with-go