Use name to identify groutine pprof label
Instead of dynamically getting the name based on the work function,
use a param string. This removes the need for reflect, but it does
mean that name config is now required. As such creating a runner
will return an error if the name is empty. This should ensure that
we don't have issues with missing labels.

The goal is to enable better development and debugging of live
production servers, so adding a mandatory name in certain locations
seems worthy of the effort.