feat: restart delay function
Allow the runner to back-off workers depending on the number of
restarts. If a given worker in the runner is starting to thrash,
we want to allow the backing-off a worker. Because the runner and
its workers are longed live (or at least expected to be) the
attempts could increase over time, which could lead to extended
back-off times. To help solve that, a restarted time is sent, which
will be a zero time on first start, but will allow the method to
be able to account of the time durring the restart attempts.
Additionally, the last error the worker died with will be useful if
there are certain errors that need to be handled differently in
terms of restarts.

A follow on PR will be added to expose the metrics of the restarts
of a given worker, by name, to identify potential issues.