feat: remove resource log due to data race
The resource/access log (inconsistent depending on variable name)
was never guarded by a mutex, so a get request and a write of a
snapshot would cause a data race.

Whilst investigating this it was noticed that you don't need the
resource/access log as nobody was reading it. It was just being
saved into a slice. The slice was never cleared, so on long running
controllers, this would just consume more and more memory.
Considering that when controllers or models restart the log would
grow considerably (albeit slowly) - growth is growth.