refactor: remove poor logging from worker runner
This commit removes a debug log saying a worker is started in the runner
because we already have a log for this at the info level. You end up
just getting double the logs for no reason.

The second fixes the killing log of the runner it self. Before this it
was loggin the pointer of the runner which offers no context. This
change logs the runner name instead.