refactor: logging of worker runner to be more consistent
This commit takes the last commit a step further by refactoring all of
the logging with in the worker runner. This change now makes it so that
the main handlers for individual channel events are the only thing that
will log out at a level of Info or greater. This is the correct place to
perform this as it is the main handling logic where decisions are made.

For other code that is responsible for getting the events sent they now
log out at debug and trace. This allows developers to confirm code
execution but does not clutter the normal logging view of a Juju
deployment.